endpoint: move `getK8sPodLabels` to `labels.go`
Signed-off by: Ian Vernon <ian@cilium.io>
